More editor refactor (#60)
It turns out™ the previous refactor is not good enough. With that, it is
easier to make mistake that keep mounting/unmounting Map.ts. It is
slow(need to initialize mapbox and deckgl) and also it doens't do clean
up correctly.

I believe this PR does make things better. And now the only sad part is
the `Map.ts`. but it is very hard to make mistake and other componets
won't be affected.