add globe projection [BETA] (#100)
Add support of globe projection.
Why beta? This is a known bug[1] causing we render things accross
meridian wrongly
in globe projection. We could workaround it by not rendering things
across meridian (e.g. having two canvas sources, one for each side). But
for now, I'm too lazy to do it, I'd rather wait for upstream's fix or do
it later and in a different PR.

[1] https://github.com/mapbox/mapbox-gl-js/issues/12758

---------

Co-authored-by: Zed(Zijun) Chen <zed@zijun.dev>